Reorganization of Files
In the previous iteration, all eight analytical functions were in the file, random_walk.py. As a result, it's extremely inconvenient to scroll back and forth to see how and what which each function does specifically. By reorganizing each function in its separate file, we're able to call that function with "from...import..." into the main file, random_walk.py.